CluStream: Offline clustering for CluStream
Moves the kMeans step to a separate function called offline_processing.
The function is called the first time a prediction is made for a timestamp. This guarantees an up-to-date kMeans clustering when predicting (which also resolves the issue of the mismatch between mc.center and mc_centers).
For compatability reasons the function is still called based on time_gap.